int to float

+++ Road.cpp
@@ -1,6 +1,6 @@
 #include "Road.h"
 
-Road::Road(Point *pr, Point *pl, string &road_name, int length)
+Road::Road(Point *pr, Point *pl, string &road_name, float length)
 {
     this->pl = pl; 
     this->pr = pr;
